// Calla-Site/src/configuration.d.ts
export declare const JITSI_HOST = "tele.calla.chat";
export declare const JVB_HOST = "tele.calla.chat";
export declare const JVB_MUC: string;
export declare const rooms: {
    value: string;
    text: string;
}[];
export declare const emojiFont: {
    fontFamily: string;
    fontSize: number;
};
export declare const audioClips: {
    join: string;
    leave: string;
};